fix : make possession Page as a component
